requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
django-redis
psycopg2
psycopg2-binary
redis
django-environ
python-decouple
celery
reportlab
openpyxl